requirements.txt
streamlit
google-genai
yt-dlp